move row prefix calculation into PrefixRowFilter
